Frontend: Pass VEX.L encoding along with decoded instruction
Vector widths are specified by VEX.L (256-bit if set, 128-bit or scalar
if not). Ideally we'd handle this all up front, but depending on the
encoding of the instruction, the size of the scalar being handled can
differ on a per-instruction basis, so this will need to be handled in
the opcode implementations where necessary.